domain,repository update

+++ src/main/java/com/project/shop/item/domain/Category.java
@@ -12,7 +12,7 @@ public class Category {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "categoryId")
-    private int categoryId;     //카테고리 번호
+    private long categoryId;     //카테고리 번호
     @Column(name = "categoryName", nullable = false)
     private String categoryName;    //카테고리 이름
     @Column(name = "brandName", nullable = false)
