+++ src/main/java/com/project/shop/item/dto/response/ReviewResponse.java
@@ -17,7 +17,7 @@ public class ReviewResponse {
     private String categoryName;    //카테고리명
     private String brandName;       //브랜드명
     private String itemName;    //상품명
-    private int userId; //회원id
+    private long userId; //회원id
     private String title;   //리뷰 제목
     private String content; //리뷰 내용
     private int star;    //별점
@@ -25,6 +25,7 @@ public class ReviewResponse {
     private LocalDateTime insertDate;   //리뷰 등록일
 
 
+    //review Entity를 ReviewResponse로 만들기
     public static ReviewResponse fromEntity(Review review){
         var list = review.getReviewImgList()
                 .stream().map(x -> new ReviewImgResponse(x.getImgUrl()))
