+++ src/main/java/com/project/shop/order/domain/PayCancle.java
@@ -10,9 +10,9 @@ public class PayCancle {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "payCancleId")
-    private int payCancleId;     //결제취소번호
+    private long payCancleId;     //결제취소번호
     @Column(name = "orderId", nullable = false)
-    private int orderId;     //주문 번호
+    private long orderId;     //주문 번호
     @Column(name = "usedPoint", nullable = false)
     private int usedPoint;     //취소 포인트
     @Column(name = "payCompany", nullable = false)
